Implemented pixel to world-position.

+++ common/RenderPipeline.h
@@ -22,7 +22,7 @@ public:
 
     RenderPipeline();
     void ClearBuffers();
-    float * RayTracePixel(int x, int y, bool geometry);
+    uint8_t * RayTracePixel(int x, int y, bool geometry);
     
     GLuint frameBuffer;
     GLuint depthBuffer;
